backend/requirements.txt
fastapi==0.123.9
uvicorn[standard]==0.24.0
aiogram==3.13.1
sqlalchemy==2.0.44
asyncpg==0.31.0
redis[hiredis]==5.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
